Add geoJson layer to map and a getBounds method
With the GeoJson tag you can add easily a GeoJson-Layer to the map.
As it is not possible to use the fitBounds method at the GeoJson-Layer,
the fitBounds method has to be applied to the map-component.
A getBounds method in the main component, returning the rectangle data does the job well.